backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
praw==7.7.1
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
cors==1.0.1
fastapi-cors==0.0.6
openai==1.55.0 